# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: JobPing API
  description: |
    JobPing is a sophisticated job matching platform for early-career professionals.
    This API provides comprehensive job matching, user management, and analytics capabilities.
    
    ## Key Features
    - AI-powered job matching with GPT-4
    - Rule-based fallback matching
    - Real-time job scraping from multiple sources
    - User preference management
    - Email automation and notifications
    - Payment processing with Stripe
    - Comprehensive analytics and monitoring
    
    ## Authentication
    Most endpoints require authentication via JWT tokens or API keys.
    Public endpoints are clearly marked.
    
    ## Rate Limiting
    API requests are rate-limited per endpoint to ensure fair usage:
    - Public endpoints: 20 requests/minute
    - Authenticated endpoints: 100 requests/minute
    - Resource-intensive endpoints: 2-5 requests/minute
    
    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
    See the Error Response schema for details.
  version: 1.0.0
  contact:
    name: JobPing Support
    email: support@jobping.ai
    url: https://jobping.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://jobping.ai/api
    description: Production server
  - url: https://staging.jobping.ai/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /webhook-tally:
    post:
      tags:
        - Webhooks
      summary: Tally Form Webhook
      description: |
        Receives form submissions from Tally forms for user registration.
        This is a public endpoint that processes new user signups.
        
        **Rate Limit**: 10 requests per minute
      operationId: webhookTally
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TallyWebhook'
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  userId:
                    type: string
                    example: "user_123"
                  email:
                    type: string
                    example: "user@example.com"
                  matchesFound:
                    type: integer
                    example: 12
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for system authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
  schemas:
...
    TallyWebhook:
      type: object
      properties:
        eventId:
          type: string
          example: "evt_1234567890"
        eventType:
          type: string
          example: "FORM_RESPONSE"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        formId:
          type: string
          example: "form_1234567890"
        responseId:
          type: string
          example: "resp_1234567890"
        data:
          type: object
          properties:
            fields:
              type: array
              items:
                type: object
                properties:
                  key:
                    type: string
                    example: "email"
                  label:
                    type: string
                    example: "Email Address"
                  type:
                    type: string
                    example: "EMAIL"
                  value:
                    oneOf:
                      - type: string
                      - type: array
                        items:
                          type: string
                      - type: "null"
                    example: "user@example.com"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Invalid email format"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        requestId:
          type: string
          example: "req_1234567890"
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Bad Request"
            message: "Invalid request parameters"
            code: "BAD_REQUEST"
            timestamp: "2024-01-15T10:30:00Z"
...
    RateLimited:
      description: Too many requests - rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  retryAfter:
                    type: integer
                    example: 60
          example:
            error: "Rate Limited"
            message: "Too many requests"
            code: "RATE_LIMITED"
            retryAfter: 60
            timestamp: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            code: "INTERNAL_ERROR"
            timestamp: "2024-01-15T10:30:00Z"